+++ src/server/physics/event/ig_ENTITY_EventTrigger.ts
@@ -27,8 +27,11 @@ prestart(() => {
         update() {
             if (!isPhysics(multi.server)) return this.parent()
 
-            const setBy = findSetByEntityByVars(this.startCondition.vars)
-            if (setBy) setNextSetBy(setBy)
+            const entities = findSetByEntityByVars(this.startCondition.vars)
+            if (entities.length > 0) {
+                const setBy = entities.find(e => e instanceof dummy.DummyPlayer) ?? entities[0]
+                setNextSetBy(setBy)
+            }
 
             if (!this.forceRunOnMap) setEventNextTriggeredBy(this.startCondition)